# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple ETMS API's for managing expenses by own or by others with groups",
        "title": "ETMS API",
        "termsOfService": "https://swagger.io/terms",
        "contact": {
            "name": "Praveen"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/groups/accept_invitation": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Accept invitation to group member with the provided group name and email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "Accept invitation to group member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Accept Inviation",
                        "name": "acceptInvitation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/groups.AcceptInvitationStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/endpoints.AcceptInvitationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/endpoints.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/endpoints.InternalServerResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/groups/creategroup": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new group with the provided group name to manage the expenses between the tenets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "Create a new group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Group",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/groups.Group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/endpoints.CreatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/endpoints.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/endpoints.InternalServerResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/groups/get_users": {
            "post": {
                "description": "get all the users present in the group by the group name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "Get Users Present in the group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group name",
                        "name": "group_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/endpoints.GetUsersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/endpoints.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/endpoints.InternalServerResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/groups/invite": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update group name with the new group name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "Update the group name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Invite User",
                        "name": "invite",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/groups.Invitation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/endpoints.InviteGroupMembersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/endpoints.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/endpoints.InternalServerResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/groups/remove_group_member": {
            "put": {
                "description": "Remove member from the group based on the email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "Remove member from the group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Remove Group Member",
                        "name": "removeMember",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/groups.AcceptInvitationStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group updated successfully",
                        "schema": {
                            "$ref": "#/definitions/groups.Group"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/endpoints.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/endpoints.InternalServerResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/groups/update_group_name": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update group name with the new group name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "Update the group name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Update Group",
                        "name": "updateGroup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/groups.UpdateGroupStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/endpoints.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/endpoints.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/endpoints.InternalServerResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "endpoints.AcceptInvitationResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "endpoints.BadRequestResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/endpoints.ErrorMessage"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "endpoints.CreatedResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "endpoints.GetUsersResponse": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "endpoints.InternalServerResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/endpoints.ErrorMessage"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "endpoints.InviteGroupMembersResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "non_existing_users": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "total_no_existing_users": {
                    "type": "integer"
                }
            }
        },
        "endpoints.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "groups.AcceptInvitationStruct": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "groupname": {
                    "type": "string"
                }
            }
        },
        "groups.Group": {
            "type": "object",
            "required": [
                "createdBy",
                "groupName",
                "type",
                "users"
            ],
            "properties": {
                "createdBy": {
                    "type": "string"
                },
                "groupName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "groups.Invitation": {
            "type": "object",
            "properties": {
                "groupname": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "groups.UpdateGroupStruct": {
            "type": "object",
            "properties": {
                "groupname": {
                    "type": "string"
                },
                "new_group_name": {
                    "type": "string"
                }
            }
        },
...
    }
}